Fixes processing XLSX with Data Model .data part
The value of Magick.formats depend on the installed ImageMagick. On
EngineYard, .data was not a format. The ImageMagick currently on AWS
instances return .data as one of its formats. This is causing the error
"Magick::ImageMagickError - no decode delegate for this image format"
when an XLSX file includes a Data Model part (\xl\model\item.data)

If we are unable to decode the blob, we now fallback to UnknownPart
which should match the previous behavior on EngineYard.